fixed doubles

+++ HW5 Sorting/Sort.h
@@ -5,8 +5,8 @@ template<class ItemType>
 class SortInterface {
 public:
     virtual void sort(vector<ItemType>& data) = 0;
-    virtual int getComparisons() = 0;
-    virtual int getSwaps() = 0;
+    virtual double getComparisons() = 0;
+    virtual double getSwaps() = 0;
     virtual double getRuntime() = 0;
     virtual void resetInformation() = 0;
     virtual ~SortInterface() {};
