+++ HW5 Sorting/vectorBubbleSort.h
@@ -15,9 +15,9 @@ private:
 public:
 	vectorBubbleSort() //init variables in constructor
 	{
-		number_of_comparisons = 0;
-		number_of_swaps = 0;
-		run_time = 0;
+		number_of_comparisons = 0.0;
+		number_of_swaps = 0.0;
+		run_time = 0.0;
 	}
 
 	void sort(vector<ItemType>& data)
@@ -41,9 +41,9 @@ public:
 		run_time = double(end_time - start_time);
 	}
 
-	int getComparisons() { return number_of_comparisons; }
+	double getComparisons() { return number_of_comparisons; }
 
-	int getSwaps() { return number_of_swaps; }
+	double getSwaps() { return number_of_swaps; }
 
 	double getRuntime() { return run_time; }
